9.12直播课-分页查询逻辑

+++ src/main/java/com/hogwartsmini/demo/entity/HogwartsTestJenkins.java
@@ -52,7 +52,7 @@ public class HogwartsTestJenkins extends BaseEntityNew {
      * 命令运行的测试用例类型  1 文本 2 文件
      */
     @Column(name = "command_run_case_type")
-    private Byte commandRunCaseType;
+    private Integer commandRunCaseType;
 
     /**
      * 备注
